[feat] : search 패키지 별도 분리
