+++ core/src/main/java/dev/handsup/auction/service/AuctionService.java
@@ -2,22 +2,15 @@ package dev.handsup.auction.service;
 
 import static dev.handsup.auction.exception.AuctionErrorCode.*;
 
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import dev.handsup.auction.domain.Auction;
 import dev.handsup.auction.domain.product.product_category.ProductCategory;
 import dev.handsup.auction.dto.mapper.AuctionMapper;
-import dev.handsup.auction.dto.request.AuctionSearchCondition;
 import dev.handsup.auction.dto.request.RegisterAuctionRequest;
 import dev.handsup.auction.dto.response.AuctionResponse;
-import dev.handsup.auction.repository.auction.AuctionQueryRepository;
 import dev.handsup.auction.repository.auction.AuctionRepository;
 import dev.handsup.auction.repository.product.ProductCategoryRepository;
-import dev.handsup.auction.repository.search.RedisSearchRepository;
-import dev.handsup.common.dto.PageResponse;
 import dev.handsup.common.exception.NotFoundException;
 import lombok.RequiredArgsConstructor;
 
@@ -27,8 +20,7 @@ public class AuctionService {
 
 	private final AuctionRepository auctionRepository;
 	private final ProductCategoryRepository productCategoryRepository;
-	private final AuctionQueryRepository auctionQueryRepository;
-	private final RedisSearchRepository redisSearchRepository;
+
 
 	public Auction getAuction(Long auctionId) {
 		return auctionRepository.findById(auctionId)
@@ -39,16 +31,6 @@ public class AuctionService {
 		ProductCategory productCategory = findProductCategoryEntity(request);
 		Auction auction = AuctionMapper.toAuction(request, productCategory);
 		return AuctionMapper.toAuctionResponse(auctionRepository.save(auction));
-	}
-
-	@Transactional(readOnly = true)
-	public PageResponse<AuctionResponse> searchAuctions(AuctionSearchCondition condition, Pageable pageable) {
-		Slice<AuctionResponse> auctionResponsePage = auctionQueryRepository
-			.searchAuctions(condition, pageable)
-			.map(AuctionMapper::toAuctionResponse);
-		redisSearchRepository.increaseSearchCount(condition.keyword());
-
-		return AuctionMapper.toAuctionPageResponse(auctionResponsePage);
 	}
 
 	private ProductCategory findProductCategoryEntity(RegisterAuctionRequest request) {
