+++ core/src/main/java/dev/handsup/search/dto/PopularKeywordResponse.java
@@ -1,4 +1,4 @@
-package dev.handsup.auction.dto.response;
+package dev.handsup.search.dto;
 
 import java.util.List;
