+++ core/src/test/java/dev/handsup/auction/service/AuctionServiceTest.java
@@ -6,32 +6,26 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.BDDMockito.*;
 
 import java.time.LocalDate;
-import java.util.List;
 import java.util.Optional;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.SliceImpl;
 
 import dev.handsup.auction.domain.Auction;
 import dev.handsup.auction.domain.auction_field.PurchaseTime;
 import dev.handsup.auction.domain.auction_field.TradeMethod;
 import dev.handsup.auction.domain.product.ProductStatus;
 import dev.handsup.auction.domain.product.product_category.ProductCategory;
-import dev.handsup.auction.dto.request.AuctionSearchCondition;
 import dev.handsup.auction.dto.request.RegisterAuctionRequest;
 import dev.handsup.auction.dto.response.AuctionResponse;
 import dev.handsup.auction.repository.auction.AuctionQueryRepository;
 import dev.handsup.auction.repository.auction.AuctionRepository;
 import dev.handsup.auction.repository.product.ProductCategoryRepository;
 import dev.handsup.auction.repository.search.RedisSearchRepository;
-import dev.handsup.common.dto.PageResponse;
 import dev.handsup.common.exception.NotFoundException;
 import dev.handsup.fixture.AuctionFixture;
 import dev.handsup.fixture.ProductFixture;
@@ -120,28 +114,5 @@ class AuctionServiceTest {
 			() -> assertThat(auctionResponse.purchaseTime()).isEqualTo(registerAuctionRequest.purchaseTime()),
 			() -> assertThat(auctionResponse.productCategory()).isEqualTo(registerAuctionRequest.productCategory())
 		);
-	}
-
-	@DisplayName("[경매를 정렬, 필터링하여 검색할 수 있다.]")
-	@Test
-	void searchAuctions() {
-		//given
-		Auction auction = AuctionFixture.auction(ProductCategory.of(DIGITAL_DEVICE));
-		PageRequest pageRequest = PageRequest.of(PAGE_NUMBER, PAGE_SIZE);
-		AuctionSearchCondition condition = AuctionSearchCondition.builder()
-			.keyword("버즈")
-			.build();
-
-		given(auctionQueryRepository.searchAuctions(condition, pageRequest))
-			.willReturn(new SliceImpl<>(List.of(auction), pageRequest, true));
-
-		//when
-		PageResponse<AuctionResponse> response
-			= auctionService.searchAuctions(condition, pageRequest);
-
-		//then
-		AuctionResponse auctionResponse = response.content().get(0);
-		Assertions.assertThat(auctionResponse).isNotNull();
-		verify(redisSearchRepository).increaseSearchCount(condition.keyword());
 	}
 }
